Moved Directory logic and js to separate Controller and js file
